Fix #334 invalid regex for methods/fields Explanation:
Added this regex to the pattern for "words" in the declaration: (?:<.+>)?

This is a noncapturing group which matches anything inside angle brackets (<>), or nothing (the ? character at the end assures it only captures 0-1 instances).

Also added this regex to pattern (method only) to avoid matching return statments use srg names: (?!return)

This is a negative lookahead to check that there is no return statement just after the whitespace.

See http://regexr.com/3cl58 for evidence of the fix.