Rewrite McpCleanup.stripComments to fix comment stripping bugs
The method contained some bugs, like how asterisk in multi line comments would
not be stripped and strings ending with asterisks would cause comments to be
recognized within strings subsequent. The old version of the method was patchable,
but it would be more difficult than necessary to think about the correct behavior
as the logic for each state(if the current character was in code, a string or a comment)
was handled in several different places. Instead, the rewrite puts each state
into its own part of the code to make it easy to reason about.