Mostly finished the new MCP maven repo
Access transformers are still not handled by the core MCP system in any way, but they are already identified and addressed by the repository to avoid overlap between setups with different ATs.
Repository-generated files will be stored in `mcp/<mcversion>/<pipeline>/<athash>` under the gradle caches directory.
"Extras" files are not handled either. These are the libraries and other things that are extracted from the client and server jars before they are deobfuscated and which should be added back to the classpath to allow the game to function correctly.
Dependency passthrough from Minecraft to the artifact is not implemented, but the core system to allow it is. A bit of reference code has been left to show how it could be done.