Prevent NPE when attempting to find patches for a vanilla Minecraft dependency.
When a userdev environment requests the following as a dependency:
`minecraft("net.minecraft:server:1.13")`
Minecraft itself will not have a set of patches to apply to it's own
source, so, we need to acknowledge that it will not have it's own
additional run arguments to add onto the runs configurations like
patchers will (like depending on forge).

There are no side effects to this change except fixing dependency
resolution for vanilla jars.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>