Ensure that gradle tasks are not run during dependency resolution
As described in
https://github.com/MinecraftForge/ForgeGradle/issues/550,
running Gradle tasks from a dependency resolution can easily lead to a
deadlock.

To avoid this issue, we perform all of the gradle task-related actions -
decompiling, recompiling, etc - up front, during plugin initialization.
This is accomplished by explicitly invoking BaseRepo#getArtifact for all
of the possible 'generated' artifacts. This ensures that all of the
caches are populated by the time dependency resolution actually runs.
With all requests for 'generated' files being served from various
caches, there's no longer any need to execute any new gradle tasks
during dependency resolution.

To ensure that the dependency resolver isn't actually generating any new
files, we set a flag 'alreadyBuiltDependencies' after the initial plugin
configuration. This flag is checked whenever a file is about to be
generated due to a cache miss - if it's set, an error is logged to
indicate the bug.