Add proper names and task creation
Newer versions of Gradle forbid creating a task with the same name
as another task, and with the way that the reobf tasks are re-created
on evaluation, this leads to several failures when multiple projects
are concerned. The fix is mainly targeting Gradle 6 enforcements of
these rules, but nonetheless are available to all gradle versions
applicable thereof.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>