Fix corruption of Unicode characters on Windows systems
I had an issue where a corrupted copy of ClientHooks.java was found inside a jar file:

...\build\tmp\expandedArchives\forge-1.16.3-34.1.0_mapped_snapshot_20201028-1.16.3-sources.jar_d036fe06dff3dbdda94c44ac227acd65\net\minecraftforge\fml\client\ClientHooks.java:183: error: unmappable character (0x81) for encoding UTF-8
        return description.endsWith(":NOFML?��?��r") ? description.substring(0, description.length() - 8)+"?��?��r" : description;

Best practices for file I/O streams is always to specify the character set. I think it's reasonable to just enforce UTF-8 in all situations, so I fixed that.

To find other cases, I added forbidden-apis to the project, and fixed all the other things it found.