Add an OFFLINE_MODE to Utils class.
Added a Utils.initUtils(Project p) that can be used to read in a
Project's parameters (and other information).

Additionally, added an OFFLINE_MODE value that is populated by the
Project's `--offline` flag. The OFFLINE_MODE flag is used in several
locations through out the Util class to prevent network access when the
flag is set.