Change replaceArgs to accept Map<String, Collection<?>>
Previously, because of how multimaps worked, a multi-replacement without any
values would result in the replacement token (and the prefix) remaining in place.
By changing from a mulitmap to a map of collections, this allows for
'empty inputs', where the key still exists in the map and therefore the
replacement token and prefix are removed even if there are no values
for that key.