Lazily evaluate updated tokens map for run tasks
The updated tokens map is created by passing in the contents of the
`runtimeClasspath` configuration. However, the run tasks are configured
with the updated tokens map, which means the configuration is queried
on configuration, which is deprecated and liable to be invalid in Gradle
8 according to [1] and the warnings when running FG.

To avoid resolving the configuration at configuration time, then, the
updated tokens map is lazily created and used for configuring the run
tasks.

The `WrappedArgument` class provides a way to lazily provide a (optionally
post-processed) string, overriding the `toString()` method to return the
string instead of the default object behavior. This allows us to pass the
object directly into the various configuration methods of JavaExecSpec,
which internally store `Object`s which are lazily `toString()`ed when the
actual arguments/environment variables/system properties are needed for
execution.

As far as I can tell, this is the only place we need to do this: while
the configuration is also resolved as part of the individual IDE run
configuration generation tasks, it is resolved within the task execution
as opposed to the task configuration of the run tasks.

[1]: Gradle 7.1 User Guide, "Viewing and debugging dependencies",
subsection "Resolving unsafe configuration resolution errors"