Add explicit task ordering to prevent Gradle dep. warnings
Gradle 7 now outputs more warnings for different wrong behaviors. One
of these behaviors is when a task has an input set
indirectly to the output of another task, without any ordering or
dependency defined between them (whether via `dependsOn`, `mustRunAfter`,
or passing a `Provider` from the output to the input).

To fix this, we define some tasks to `mustRunAfter` other tasks which
produce their inputs without being connected via `dependsOn` or `Provider`s.
Through this, Gradle will ensure that the producing tasks run _before_ the
consuming tasks, but will not force the producers to run (which is what
`dependsOn` specifies).