Fix eclipse run configurations for multi-projects
Originally, the Eclipse runs generator uses the eclipse model of the
current project to retrieve the classpath, in order to make a mapping
of source set names to their corresponding output folders. This is
important so the mod code loaded by FML is the classes compiled by
Eclipse, not by Gradle.

However, there's a problem when a run config uses multiple source
sets with the same name from different projects

First, because only the current project's eclipse model is queried,
the run generator has no information on the output folders for source
sets outside the current project. Second, there is no way to retrieve
the project for a source set in order to get the eclipse model for
those projects because source sets do not carry with them their owning
project's information. Third, because the map keys are only source set
names, there is no way to retrieve the proper output folder for source
sets which share the same name but come from different projects.

To solve this, there are now two maps: one is the `allOutputs` map
which is the original map of source set names to output folders, and
the new one is `sourceSetsToOutputs`, which tracks `SourceSet`
instances (using an `IdentityHashMap`) to a corresponding
project-specific map of source set names to output folders. For
each `SourceSet` that belongs to a project, they all share the
same project-specific outputs map. This allows looking up first by
source set for the project-specific outputs map (falling back to the
original shared outputs map if not present), then by source set name.

To create the `sourceSetsToOutputs` map, all projects within the Gradle
build are iterated over and retrieved for the `eclipse` plugin if it is
present (which is usually the case when running from Eclipse), and then
retrieving the eclipse model object and classpath from there.