Add map of source sets to owning projects
The project of a source set is required to properly calculate the IDEA
output path for that source set.

Specifically, it is needed for the project's directory according to the
IdeaModel or, if that is not available, the project's directory.

This is similar to the solution present in
EclipseRunGenerator#mapModClassesToEclipse, wherein a map of source
sets to source-set-specific data is kept.