Use IDEA module info instead of project when calculating output path
The IDEA project info, represented as an IdeaProject object, only
exists on the root project. However, we want to use the module info,
represented as an IdeaModule object. This is because the Gradle project
is closer in concept to the IDEA module, whereas the IDEA project is
closer to the whole Gradle build (root project).