Move transient field assignment to getters
This lazily constructs and assigns the values of the transient fields
only when requested through the getters, rather than in the
package-private constructor.

This fixes an issue where equals and hashCode (which use a transient
field) are called on an Artifact after being deserialized, causing an
NPE as the transient fields have no value after deserialization.