[6.0] Version catalog version support (re-try) (#980)
This reapplies commit 0b1f02a52d71f29cf77ad74534b1a181e1dd7d52.

`GString` and other subclasses of `CharSequence` exists. The solution is
to run `Object#toString` on the resultant value rather than casting
it to a string. This is the same behavior as before, the only difference is
that the enforcement on `CharSequence` is removed.