Make `minecraft.mavenizer(repositories)` add to top of the list
Traditionally, maven repositories are added like this:
```groovy
repositories {
    maven minecraft.mavenizer
    maven { url = 'https://example.com/repo' }
}
```
In the case of Mavenizer, which we strongly recommend (and require in
most cases) to be at the top of the repositories list, I've updated
`minecraft.mavenizer(repositories)` to do that. You can use it like
this:
```groovy
minecraft.mavenizer repositories
// or
repositories {
    minecraft.mavenizer it
}
```
May be useful for complex setups that add to repositories early.