Create Eclipse launchers based off of Slime Launcher run configs (#1005)
These are very similar to the Eclipse run configurations generated by ForgeGradle 6. While generally still a work-in-progress in terms of resilience and feature set, they are functional and should work without issue as long as the consuming project doesn't use substitutions in resources.

I'm still brainstorming ways to get around the fact that debugging JavaExec tasks is rather limited in Eclipse Buildship at the moment. I'll need to play around a bit more with bundled run XMLs for Eclipse, and see how the debugger treats those.

- Bump embedded GradleUtils Shared to 3.3.38