Allow genEclipseRuns to be executed without the Eclipse plugin
The point of requiring the Eclipse plugin for `genEclipseRuns` was to get the accurate compiler output directory and project name as defined by the Eclipse project. However, the absense of the task itself can cause issues in various buildscript setups.

Since it is generally difficult to configure these attributes for an Eclipse project, the task will now run without the Eclipse plugin and use best-guess defaults for these values. However, a warning will be emitted to the problems reporter if any Eclipse run config generation task is run without the Eclipse plugin being loaded.